+++ app/packages/routers/__init__.py
@@ -1,11 +1,3 @@
 from .DeepFM import *
-from .DeepFM_V2 import *
-
 from .Sbert import *
 from .d2v import *
-# from app.packages.routers import *
-
-
-# from app.packages.routers.Sbert import *
-# from app.packages.routers.d2v import *
-# from app.packages.routers.DeepFM import *
