+++ requirements.txt
@@ -65,7 +65,7 @@ requests==2.31.0
 requests-oauthlib==1.3.1
 rsa==4.9
 safetensors==0.4.1
-scikit-learn==1.3.2
+scikit-learn==1.3.0
 scipy==1.11.4
 sentence-transformers==2.2.2
 sentencepiece==0.1.99
@@ -78,7 +78,6 @@ tensorboard==2.15.1
 tensorboard-data-server==0.7.2
 tensorflow==2.15.0
 tensorflow-estimator==2.15.0
-# tensorflow-intel==2.15.0
 tensorflow-io-gcs-filesystem==0.31.0
 termcolor==2.4.0
 threadpoolctl==3.2.0
